requirements.txt
discord.py
discord-ext-voice-recv
edge-tts
faster-whisper
python-dotenv
langdetect
numpy
